Remove unused parameters from cancel
cancel as a protected function is only called from one place and the processContribution value is always false.
This means that the returned processContribution value is always false - so we don't need to return
it as it is set to false already. In addition the currently returned updateResult is no longer
used so we don't need to return that either.

I will do the same changes in a separate PR for processFail