Add test demonstrating that an extraneous activity is being created
This test checks that 1 & only 1 activity is created when a membership is cancelled because the related contribution is cancelled.

It is currently failing as there is an extra 'status changed from cancelled to cancelled' activity created.

The fix is simple but not included as https://github.com/civicrm/civicrm-core/pull/18997 needs to be merged first