[REF] Duplicate possibly-used parts of createProfileContact onto Parser_Contact
The parser_Contact class should call the api. Getting to that point is
complicated because
1) it's such a mess and
2) it does a lot of work to prepare to call createProfileContact
and it's hard to unravel that.

By copying most of createProfileContact onto the class we can start to
unravel it & pare down to calling the api (note we went through
this process in the dedupe class, at least in part).

I'm pretty sure the hooks reflect 'the way the code grew up' rather than
any logical use/ requirement for the profile hook