dev/core#866, dev/core#1318 Fix failure to import checkboxes for activities
This seeks to address a long-standing issue whereby checkboxes cannot be imported onto
activities if the name and value match.

Since this code was written the create was switched from a BAO create to the
activity api create. Simply removing the handling to prepare for the BAO create
allows the api to do it correctly