Remove or hard-code variables from previously shared function
This function was separated from the shared function for cleanup.

This removes 3 variables
- isRecur (always true)
- billingId - this is elsewhere handled on the form
- online (always false)

For billing id this is handled through back office form
shared functions. The handling might be slightly different but it seems
more important that the form is internally consistent with how
it creates billing addresses (between recur & non-recur) than with other forms