Add pseudoconstant for Relationship.relationship_type_id
This adds basic pseudoconstant support for the field using name_a_b and label_a_b
(it does not support other properties like name_b_a)

For legacy purposes it preserves the work done in https://github.com/civicrm/civicrm-core/pull/11853
to return a list like [1_a_b => Child of, 1_b_a => Parent of] if the "is_form" flag is set.
This was very nonstandard but is used to support in-place editing of the relationship type list
on the relationship form.