[REF] Remove unreachable code
This code was copied from the previously shared function & supports some really weird stuff for the contribution form.

I tested & was unable to reach this from the back office membership form - note it is only reachable
when processing recurring which in turn relies on the membership type being configured and the
processor. It seems that recurring is not available with price sets