Stop using refresh_date in civicrm_group table
I was looking to add indexes to civicrm_group.cache_date & civicrm_group.refresh_date - but I couldn't figure out why
the latter exists / is used. I went through the places it's used and it is simply a calculated version of
cache_date + smartGroupCacheTime and I can't see any value.

I can follow up on this with a removal if no-one else can see the point of it