dev/core#527 use public title for on behalf profile and confirm/thank you pages
This is a partial reviewer's commit of https://github.com/civicrm/civicrm-core/pull/19144

That PR is stale. I tested one flow & added in the 2 changes from there that I could see in my test
& which I was comfortable would not affect anything else. The flow I tested was a contribution
page flow with an on behalf profile. I did not check the email as that would need
a test anyway