CRM_Utils_Cache - Extract serialization trait (APCcache)
Note that this change technically changes the storage format used by
APCcache, but it still supports reading the old format (which should
head-off any considerations with upgrade-safety).

This lays the ground-work for adding encryption for cached data. The
encrypt/decrypt mechanism will need a single point for at which to
encrypt the serialized-plaintext-value.