Accept PHP memory_limit of -1 as meeting requirement
A PHP memory_limit of '-1' is specified as 'unlimited'.

Before the change the checkmemory() function generates a Warning if the limit is set to -1.  This shows up as a Warning in the Drupal Status Report, but more seriously also prevents the Drupal database update process (mydomain.com/update.php) from running.

After the change the value '-1' is accepted as meeting the requirement.