distmaker - Don't require dummy config file for building Joomla
BEFORE: To run distmaker and produce Joomla ZIPs, you need to create a dummy settings file.
For example, this is how `cividist` does it:

https://github.com/civicrm/civicrm-buildkit/blob/v19.07.0/bin/cividist#L119-L121

It seems that is just a long-winded way of setting up the Civi classloader. (It wouldn't
make sense to use any other Civi services this is a pre-bootstrap environment.)

AFTER: The dummy file is neither needed nor used. We just get the classloader directly.