dev/core#2303 Extract function to load the messageContent for a template
This is a fairly straightforward extraction, seeking to break down the very-busy sendtemplate function
into the sub-functions within it. Essentially we load the template content, we swap out the tokens
and we send it.

Note that I have added a parameter to the output sent to alterMailContent
of 'workflow_name' which is more accurate - I will update the docs