Bridge search displays and afforms
This allows an afform af-fieldset to function without a containing af-form or reference to af-entity
Instead it can directly contain an api-entities tag to assist filling field metadata.
The search extension then adds that tag based on the search parameters.
This also eliminates the need to pass in filters to the search display directly (though it still supports that),
as it will infer them from the presence of an af-fieldset.