AfformAdmin - use api-callback for loading metadata
This switches from pre-loading all field data, to a more incremental approach of only loading the data needed by the form being edited,
then loading more metadata if additional entities or blocks are added to the form.

The Afform.loadAdminData api action is added to facilitate loading metadata for different types of afforms,
and incrementally loading metadata for new entities & blocks as they are added to an afform being edited.