APIExplorer - Split "Rest" in two ("Rest" vs "curl")
We're serving a dual purpose of educating generally about the REST request,
and providing an example that is runnable.

However, we're not doing either well if the REST example is simultaneously
incorrect and an obscure block of text.

This splits it into two examples, one emphasizing general structure
and the other emphasizing precision/correctness.