Remove campaign_id pseudoconstants
Pseudoconstants are intended for option lists small enough to be loaded into memory,
whereas FKs are for options of any size.
It is possible to have thousands of campaigns, and some sites do,
which slows or breaks some UIs (first noticed in the Search Kit bulk update action).

All these fields have both an FK and a pseudoconstant declared, which is redundant;
both are not needed, and the pseudoconstant was unnecessarily hogging memory for sites with many campaigns.