Set api default for contributionRecur to Pending
This has a DB default of 'Completed' which is not right as far as I'm concerned - the right
default is 'Pending'. We could update the DB default or add defaults on both
v3 & v4 api.

I think we would need to advise of this change via change log, release notes, dev-digest
but I doubt it would be that controversial since contributions need to be
add to to contribution recur after it is created & if enough
contributions are added to complete it it would be updated to Completed