[REF] Simplify is_email_receipt handling
This removes a chunk of confusing handling from the previously shared code.

Philosophically we have a lot of arrays of data being passed around. The difference
between these arrays is unclear & the array that appears to exist specifically to
how what was actually submitted (formValues) is edited rather than left intact.

Where we are not dealing with a calcuated value, but rather referring to
what was submitted the quickform function getSubmittedValue() seems
like a good pick. It's clear we are looking up the submitted value not
some calculated value & when code is moved around we know this
value is not dependent on where it is sitting in the code.

The function returns 1 or NULL - hence I cast to boolean where I'm not
sure NULL is good enough