[REF] Remove unused pcp & soft credit handling from previously shared code
the ProcessContribution function has been split out from a shared function used by front end forms
and I'm pretty sure it was only relevant for those other forms. The only code that calls this function
is when a payment processor is used in recurring mode to pay for a membership.

I tested doing this and selected 'record payment from a different contact' without these
lines & it created the soft credit (using the code elsewhere in the function that
would be used if it were not recurring or not credit card payment)