[REF] Complete order
Unset contribution once we have finished with it.

This patch does very little but it takes a bit of searching to check the handful of places
that call completeOrder and check that none of them use the contribution
object after this function call. With this merged it will be possible to
cleanup the references to the contribution in the repeattransaction
code safe in the knowledge that the object is not needed after that
function (or within it from my reading). It seems likely we can switch to
passing in contributionID rather than the whole object.