Convert token processing to use token processor for contacts
Note this will test fail on 3 things that worked with the other. The way in
which we load tokens means that some values have been exposed that
possibly would be better not exposed - but we can either remove them
from advertised tokens, fix the tokenCompat class to support them
or possibly deprecate by removing from the UI display but
continuing to handle. One notable difference is the contact_reference
entity id.