Extract 2 tests from AfformRoutingTest to mockPublicForm. Make example work.
These two cases rely on setting up the same fixture. It's easier to try out/debug if
the fixture is live/available as a matter of course.

Tangentially, if you try to use this form (whether as anonymous or as admin, it doesn't -- e.g.
because it doesn't have a "Submit" button. So this updates it.