Joomla - Allow E2E unit-test. Use Joomla Session API (and fake it when necessary).
Overview
--------

The goal of this work is to allow running some end-to-end tests on Joomla
with PHPUnit+cv. Specifically, it's aimed at the `authx` test.

The effect and approach involves realigning APIs -- ensuring that the Joomla
integration uses the Joomla Session APIs instead of the PHP Session APIs.

Before
------

* When running authx E2E test on Joomla, the phpunit process emits warnings
  and/or failures about session management.

* If Civi needs to request that a session be started or ended, it uses the
  PHP Session APIs. This *mostly* lines up with what Joomla does (since its
  default behavior is to wrap PHP Session APIs).

* There are a few use-cases based on `loadBootstrap()` -- namely, CLI (`cv`,
  `phpunit`, `bin/cli.php`) and extern (`extern/rest.php`, `extern/url.php`,
  etc).  In all these cases, it may need to set a user -- and it
  unnecessarily starts a standard PHP session.

After
-----

* When running authx E2E test on Joomla, the phpunit process has no
  warnings or failures about session management.

* If Civi needs to request that a session be started or ended, it uses
  Joomla Session APIs.  For default scenarios, this still passes through to
  the PHP Session API.  Importantly, if there's an alternative
  session-handler, then it will be respected.

* The use-cases based on `loadBootstrap()` (CLI + extern) now use an
  ephemeral session with fake data storage. The technique is inspired
  by the technique used in Joomla's upstream unit-testing.

Comments
--------

You might ask -- if CLI and extern scripts do not merit session-handling,
then why bother with the fake/ephemeral session object?  Wouldn't it be
cleaner to omit all mention of session-handling?  The main reason:
user-integration.  Any services that involve the "current user" will check
the session.  Thus, even if the process is substantively
stateless/sessionless, it must still have 'session' object for identifying
the current user.