Remove unreachable code
Term is always set at this point unless we are dealing with a priceset - which
this handling appears to be for, ergo term is always set.

We also know that this code is unreachable as in the 6 years it has been in the
codebase no-one has reported the bug in it. It uses total_amount not total_amount_numerci
in the first part of the calculation - this would fail when the decimal separator
is not a decimal