Afform - Support embedding afforms in WP shortcodes
This implements the necessary hooks & callbacks for picking an afform from the Civi shortcode popup,
and displaying the afform when viewing a single WP post/page or list of posts.

Note that unlike traditional Civi shortcodes, afform does not rely on invoking a civi page request,
but simply outputs the AngularJS directive.