Enable & do single field metadata conversion for defaultContactCountry
This builds on https://github.com/civicrm/civicrm-core/pull/19629

Our intention is that we should be building functionality into this
form that is actually used in core. It's a bit tricky to swap out
all fields at once but this makes them individually swap-outable.

Only defaultContactCountry is swapped out. I looked a little at the
help_text issue - the text actually does work if you
add it since it's already in the hlp file but I decided this
broke consistency less (using description) as other fields
on the form do that and look similar. Also, when the text is
really just a bit of text (not a long help text) it makes sense.

(Otherwise I'd need to trawl through resolving the bigger issue
and I didn't want to bring that into scope)