Export UI - Only present "save mapping" button to permissioned users
The button was showing for all users which would result in confusion when it didn't work because they lacked permission.
Currently the APIv3 Mapping.create api requires "administer CiviCRM." I don't know if that's right or if there's another
permission it ought to be checking, but at least this brings the UI in alignment with what's currently allowed.