dev/release#16 - Make incremental upgrade steps optional
Often, when setting up the initial beta and when setting up the final stable
release, there are no upgrade steps.  However, we're required to create the
files anyway to ensure that the upgrader continues to work normally.

As a byproduct, we create a large number of empty incremental steps.  When
running the web-based upgrade, each incremental step requires 3x AJAX calls.
This adds up.

Before
-----------

The revision sequence -- and the final DB version-number -- are strictly
determined by the list of `X.Y.Z.mysql.tpl` files. If you omit a file,
then it will not execute corresponding PHP code, and it will not set
the final version in DB.

After
-----------

The revision sequence is more forgiving -- it will recognize any revision
number that is defined in either `X.Y.Z.mysql.tpl` or `function upgrade_X_Y_Z()`.

Additionally, it is not necessary for the final version number to appear in
the incremental revision list.

This patch enables a slightly different workflow for developing updates:

1. Set the new version in `xml/version.xml`.

2. If (and only if) you need some patch-level incremental work, then create
   a file `X.Y.Z.mysql.tpl` or create a function `function upgrade_X_Y_Z()`
   (or both).