[REF] Cleanup code to determine financial_type_id
Financial type id is really simple on this form - it's either required or
it can be determined form the price set.

However, the code passes the financial_type_id from array to array, calculating
it in more than one place which rather hides the underlying simplicity of it.

This retrieves it using a function that does the same from anywhere in the code.

Note that if someone tried to add it before this->order is built it would
hard fail & kill a bunch of tests. this->order is built at the start of the
submit routine