[Ref] Remove copy and paste overkill
This is a previously shared function that was copied and pasted back. However in the
context of this form they don't make sense because

1) It doesn't make sense to try to maintain the integrity (via the transaction)
of the contribution & recurring contribution over & above the other
items on the form - without the memberships they don't actually have integrity
and this might have made a little more sense front end in terms
of ensuring they were committed if a payment is made but
we can expect a back office user to observe an error

2) setting the contribution id on the form here doesn't make sense as
this line is only hit for credit-card recurring contributions and
either we are doing it elsewhere anyway or we should do it nowhere - there
is no logic for 'only do this in this one obscure flow' that makes
sense back office. Front office this was likely added for an otherwise missed
scenario