Reorganize contact summary template to support non-ajax tabs
There was some non-functional (due to a typo) code in TabHeader.tpl to allow tabs to load from a .tpl instead of ajax.
This fixes that typo and copies the funcionality into the contact summary tabs as well.
The functionality is demonstrated here by moving the main tab into a sub-tempate.