Expose the receipt action for standalone or searchkit use
Adds controller for id based tasks

Supports a url like https://dmaster.localhost:32353/civicrm/contribute/task?task=receipt&id=40,41

Note the pdf option doesn't work in pop-up mode - to discuss

Also note this exposes it as an action on the contribution results. That's optional. I think it has
been requested / would be useful but it's mostly included to make the page accessible for testing
and to work through the issue around filtering - the action only works on some statuses