Fix Payment.create to update (recalculate) contribution fee_amount
This recalculates the fee amount when a payment is received. Our data integrity expects the fee
amount on the contribution to equal the sum of the fee amount on the payments. However,
it's possible a pending contribution has an 'estimated' fee amount so we can't just add on
the fee_amount and assume it will be right