dev/core#2505 Fix formatLocaleNumeric
With the back & forth over formatting this function lost it's original intent and became unused.

I think this gives back the original intent - ie provide a formatted number
with no currency symbol that reflects the locale.

I think this would be usable for https://lab.civicrm.org/dev/core/-/issues/2505
since formatting numbers should be more or less the same as formatting currencies
when only the number is being formatted.