community/feature-request#12 Separate (some) logging to different services
On reading through https://lab.civicrm.org/community/feature-request/-/issues/12#note_11253 and
thinking about some needs we have it occured to me that

'Filters: We only need to filter based on a threshold using the RFC-5424 severity (error, warning, notice, etc). No other filters are needed.' is not really correct from our point of view and
what we want is to log debug from some processes to our syslog & leave other stuff unchanged. In our case the
processes are from our own custom code but I'm pretty sure some IPN classes are already configured to
log to a different file so I was thinking about that sort of use case when I did this quick proposal.

Effectively the idea is that when we log things we could log to the default (psr_log) or to a different
log which would, by default, do exactly the same thing - but which could be selectively changed
behaviour-wise in an extension. I pulled out the things that seemed most obviously 'categorisable'
to me.

Note having read through the issue & thought about it I am pretty sure I can meet our
immediate need using a patched version of Dave's monolog extension (in particular I'm concerned that it has
weight gain as a known side effect so I want to try to fix that and also I want to alter the destination of our
custom logger to go to syslog, not of the main psr_log) - so this is primarily to add to the discussion & unless there is some fairly
direct path to it being merged I will close it shortly