Use order api when creating a recurring membership from the Membership backoffice form
This removes the need for some 'magic' code from the membership BAO that was really only
there to support the fact that this code was doing some particularly convoluted
manoevering in order to share code with the front end form (since unshared).

Not this adds a feature Matt requested - the created membership id is returned. This
is only done for memberships at the moment but could be other entities too as
test cover is added (the membership tests fail without the change in this
PR so it has cover in the context it is added